In the Swift API, deprecate `a.dot(b)` and `?` to `matmul(a, b)` to accurately reflect the operator?s mathematical properties and make it familiar to TensorFlow users. Currently the deprecation is a warning - when we update tensorflow/swift-models, I'll start another CL to remove it completely.
Previously `dot` was chosen over `matmul` because of naming convention concerns (acronyms aren?t common in Swift) and that we wanted to make it short (so full names like `a.matrixMultiplied(by: b)` isn?t acceptable). Beyond these concerns, `matmul` is really a word of art and thus should be preferred.

The ? operator often denotes outer product and Kronecker product. So it's removed, too.

PiperOrigin-RevId: 198920621